Select instead of autocomplete controls
Switched away from jQueryUI autocomplete-type controls in favor of
regular select controls, which are better for showing a human-readable
option while storing a code value. Refactored and renamed the Cold
Fusion page that generates JSON from query results. Added a heuristic
algorithm to it for estimating size of query results and throttling the
query accordingly.